refactor: use tanstack query on explorer
